function/f1-radio-analyzer/requirements.txt
functions-framework==3.10.1
google-cloud-bigquery==3.41.0
google-cloud-logging==3.15.0
google-genai==1.73.1
python-dotenv==1.2.2
